Optimize factorial computation

+++ app/src/test/java/com/github/dzirbel/battlecompanion/core/MathTest.kt
@@ -8,45 +8,45 @@ class MathTest {
 
     @Test
     fun testFactorial() {
-        assertEquals(1.toBigInteger(), factorial(1))
-        assertEquals(2.toBigInteger(), factorial(2))
-        assertEquals(6.toBigInteger(), factorial(3))
-        assertEquals(24.toBigInteger(), factorial(4))
-        assertEquals(120.toBigInteger(), factorial(5))
-        assertEquals(720.toBigInteger(), factorial(6))
-        assertEquals(5040.toBigInteger(), factorial(7))
+        assertEquals(1, factorial(1))
+        assertEquals(2, factorial(2))
+        assertEquals(6, factorial(3))
+        assertEquals(24, factorial(4))
+        assertEquals(120, factorial(5))
+        assertEquals(720, factorial(6))
+        assertEquals(5040, factorial(7))
     }
 
     @Test
     fun testFactorialFrom() {
-        assertEquals(2.toBigInteger(), factorialFrom(n = 2, k = 0))
-        assertEquals(2.toBigInteger(), factorialFrom(n = 2, k = 1))
+        assertEquals(2, factorialFrom(n = 2, k = 0))
+        assertEquals(2, factorialFrom(n = 2, k = 1))
 
-        assertEquals(6.toBigInteger(), factorialFrom(n = 3, k = 0))
-        assertEquals(6.toBigInteger(), factorialFrom(n = 3, k = 1))
-        assertEquals(3.toBigInteger(), factorialFrom(n = 3, k = 2))
+        assertEquals(6, factorialFrom(n = 3, k = 0))
+        assertEquals(6, factorialFrom(n = 3, k = 1))
+        assertEquals(3, factorialFrom(n = 3, k = 2))
 
-        assertEquals(24.toBigInteger(), factorialFrom(n = 4, k = 0))
-        assertEquals(24.toBigInteger(), factorialFrom(n = 4, k = 1))
-        assertEquals(12.toBigInteger(), factorialFrom(n = 4, k = 2))
-        assertEquals(4.toBigInteger(), factorialFrom(n = 4, k = 3))
+        assertEquals(24, factorialFrom(n = 4, k = 0))
+        assertEquals(24, factorialFrom(n = 4, k = 1))
+        assertEquals(12, factorialFrom(n = 4, k = 2))
+        assertEquals(4, factorialFrom(n = 4, k = 3))
 
-        assertEquals(120.toBigInteger(), factorialFrom(n = 5, k = 0))
-        assertEquals(120.toBigInteger(), factorialFrom(n = 5, k = 1))
-        assertEquals(60.toBigInteger(), factorialFrom(n = 5, k = 2))
-        assertEquals(20.toBigInteger(), factorialFrom(n = 5, k = 3))
-        assertEquals(5.toBigInteger(), factorialFrom(n = 5, k = 4))
+        assertEquals(120, factorialFrom(n = 5, k = 0))
+        assertEquals(120, factorialFrom(n = 5, k = 1))
+        assertEquals(60, factorialFrom(n = 5, k = 2))
+        assertEquals(20, factorialFrom(n = 5, k = 3))
+        assertEquals(5, factorialFrom(n = 5, k = 4))
 
-        assertEquals(720.toBigInteger(), factorialFrom(n = 6, k = 0))
-        assertEquals(720.toBigInteger(), factorialFrom(n = 6, k = 1))
-        assertEquals(360.toBigInteger(), factorialFrom(n = 6, k = 2))
-        assertEquals(120.toBigInteger(), factorialFrom(n = 6, k = 3))
-        assertEquals(30.toBigInteger(), factorialFrom(n = 6, k = 4))
-        assertEquals(6.toBigInteger(), factorialFrom(n = 6, k = 5))
+        assertEquals(720, factorialFrom(n = 6, k = 0))
+        assertEquals(720, factorialFrom(n = 6, k = 1))
+        assertEquals(360, factorialFrom(n = 6, k = 2))
+        assertEquals(120, factorialFrom(n = 6, k = 3))
+        assertEquals(30, factorialFrom(n = 6, k = 4))
+        assertEquals(6, factorialFrom(n = 6, k = 5))
 
-        assertEquals(9900.toBigInteger(), factorialFrom(n = 100, k = 98))
+        assertEquals(9900, factorialFrom(n = 100, k = 98))
 
-        assertEquals((998 * 999 * 1000).toBigInteger(), factorialFrom(n = 1000, k = 997))
+        assertEquals((998 * 999 * 1000), factorialFrom(n = 1000, k = 997))
     }
 
     @Test
